Data.Array.Accelerate.Linear.Matrix

Description

Simple matrix operations for low-dimensional primitives

Documentation

(!*!) :: (Functor m, Foldable t, Additive t, Additive n, Num a, Box2 m t a, Box2 t n a, Box2 m n a) => Exp (m (t a)) -> Exp (t (n a)) -> Exp (m (n a)) infixl 7

Matrix product. This can compute any combination of sparse and dense multiplication.

>>> test $ lift (V2 (V3 1 2 3) (V3 4 5 6) :: M23 Int) !*! lift (V3 (V2 1 2) (V2 3 4) (V2 4 5) :: M32 Int)
V2 (V2 19 25) (V2 43 58)

(!+!) :: (Additive m, Additive n, Num a, Box2 m n a) => Exp (m (n a)) -> Exp (m (n a)) -> Exp (m (n a)) infixl 6

Entry-wise matrix addition.

>>> test $ lift (V2 (V3 1 2 3) (V3 4 5 6) :: M23 Int) !+! lift (V2 (V3 7 8 9) (V3 1 2 3) :: M23 Int)
V2 (V3 8 10 12) (V3 5 7 9)

(!-!) :: (Additive m, Additive n, Num a, Box2 m n a) => Exp (m (n a)) -> Exp (m (n a)) -> Exp (m (n a)) infixl 6

Entry-wise matrix subtraction.

>>> test $ lift (V2 (V3 1 2 3) (V3 4 5 6) :: M23 Int) !-! lift (V2 (V3 7 8 9) (V3 1 2 3) :: M23 Int)
V2 (V3 (-6) (-6) (-6)) (V3 3 3 3)

(!*) :: (Functor m, Foldable r, Additive r, Num a, Box2 m r a, Box m a) => Exp (m (r a)) -> Exp (r a) -> Exp (m a) infixl 7

Matrix * column vector

>>> test $ lift (V2 (V3 1 2 3) (V3 4 5 6) :: M23 Int) !* lift (V3 7 8 9 :: V3 Int)
V2 50 122

(*!) :: (Foldable t, Additive f, Additive t, Num a, Box t a, Box f a, Box2 t f a) => Exp (t a) -> Exp (t (f a)) -> Exp (f a) infixl 7

Row vector * matrix

>>> test $ lift (V2 1 2 :: V2 Int) *! lift (V2 (V3 3 4 5) (V3 6 7 8) :: M23 Int)
V3 15 18 21

(!!*) :: (Functor m, Functor r, Num a, Box2 m r a) => Exp (m (r a)) -> Exp a -> Exp (m (r a)) infixl 7

Matrix-scalar product

>>> test $ lift (V2 (V2 1 2) (V2 3 4) :: M22 Int) !!* 5
V2 (V2 5 10) (V2 15 20)

(*!!) :: (Functor m, Functor r, Num a, Box2 m r a) => Exp a -> Exp (m (r a)) -> Exp (m (r a)) infixl 7

Scalar-matrix product

>>> test $ 5 *!! lift (V2 (V2 1 2) (V2 3 4) :: M22 Int)
V2 (V2 5 10) (V2 15 20)

(!!/) :: (Functor m, Functor r, Floating a, Box2 m r a) => Exp (m (r a)) -> Exp a -> Exp (m (r a)) infixl 7

Matrix-scalar division

type M22 a = V2 (V2 a)

A 2x2 matrix with row-major representation

type M23 a = V2 (V3 a)

A 2x3 matrix with row-major representation

type M24 a = V2 (V4 a)

A 2x4 matrix with row-major representation

type M32 a = V3 (V2 a)

A 3x2 matrix with row-major representation

type M33 a = V3 (V3 a)

A 3x3 matrix with row-major representation

type M34 a = V3 (V4 a)

A 3x4 matrix with row-major representation

type M42 a = V4 (V2 a)

A 4x2 matrix with row-major representation

type M43 a = V4 (V3 a)

A 4x3 matrix with row-major representation

type M44 a = V4 (V4 a)

A 4x4 matrix with row-major representation

m33_to_m44 :: Num a => Exp (M33 a) -> Exp (M44 a)

Convert a 3x3 matrix to a 4x4 matrix extending it with zeros in the new row and column.

m43_to_m44 :: Num a => Exp (M43 a) -> Exp (M44 a)

Convert a 4x3 matrix to a 4x4 matrix, extending it with [ 0 0 0 1 ] column vector

det22 :: Num a => Exp (M22 a) -> Exp a

2x2 matrix determinant

det33 :: Num a => Exp (M33 a) -> Exp a

3x3 matrix determinant

det44 :: Num a => Exp (M44 a) -> Exp a

4x4 matrix determinant

inv22 :: Fractional a => Exp (M22 a) -> Exp (M22 a)

2x2 matrix inverse

inv33 :: Fractional a => Exp (M33 a) -> Exp (M33 a)

3x3 matrix inverse

inv44 :: Fractional a => Exp (M44 a) -> Exp (M44 a)

4x4 matrix inverse

identity :: forall t a. (Traversable t, Applicative t, Num a, Box2 t t a) => Exp (t (t a))

The identity matrix for any dimension vector.

>>> test $ (identity :: Exp (M44 Int))
V4 (V4 1 0 0 0) (V4 0 1 0 0) (V4 0 0 1 0) (V4 0 0 0 1)

>>> test $ (identity :: Exp (V3 (V3 Int)))
V3 (V3 1 0 0) (V3 0 1 0) (V3 0 0 1)

transpose :: (Distributive g, Functor f, Box2 f g a, Box2 g f a) => Exp (f (g a)) -> Exp (g (f a))

transpose is just an alias for distribute

>>> test $ transpose $ lift (V3 (V2 1 2) (V2 3 4) (V2 5 6) :: M32 Int)
V2 (V3 1 3 5) (V3 2 4 6)

class Trace m => Trace m where

Minimal complete definition

Nothing

Methods

trace :: (Num a, Box2 m m a) => Exp (m (m a)) -> Exp a

Compute the trace of a matrix

diagonal :: Box2 m m a => Exp (m (m a)) -> Exp (m a)

Compute the diagonal of a matrix

fromQuaternion :: forall a. Num a => Exp (Quaternion a) -> Exp (M33 a)

Build a rotation matrix from a unit Quaternion

mkTransformation :: forall a. Num a => Exp (Quaternion a) -> Exp (V3 a) -> Exp (M44 a)

Build a transformation matrix from a rotation expressed as a Quaternion and a translation vector.

mkTransformationMat :: Num a => Exp (M33 a) -> Exp (V3 a) -> Exp (M44 a)

Build a transformation matrix from a rotation matrix and a translation vector.