Data.Array.Accelerate.Linear.Plucker

Description

Plücker coordinates for lines in 3d homogeneous space.

Documentation

data Plucker a :: * -> *

Constructors

Plucker !a !a !a !a !a !a

squaredError :: forall a. Num a => Exp (Plucker a) -> Exp a

Valid Plücker coordinates p will have squaredError p == 0

That said, floating point makes a mockery of this claim, so you may want to use nearZero.

(><) :: forall a. Num a => Exp (Plucker a) -> Exp (Plucker a) -> Exp a infixl 5

This isn't the actual metric because this bilinear form gives rise to an isotropic quadratic space

plucker :: forall a. Num a => Exp (V4 a) -> Exp (V4 a) -> Exp (Plucker a)

Given a pair of points represented by homogeneous coordinates generate Plücker coordinates for the line through them, directed from the second towards the first.

plucker3D :: forall a. Num a => Exp (V3 a) -> Exp (V3 a) -> Exp (Plucker a)

Given a pair of 3D points, generate Plücker coordinates for the line through them, directed from the second towards the first.