Data.Array.Accelerate.Linear.Plucker

Description

Plücker coordinates for lines in 3d homogeneous space.

Documentation

data Plucker a :: * -> *

Constructors

Plucker ~a ~a ~a ~a ~a ~a

squaredError :: forall a. Num a => Exp (Plucker a) -> Exp a

Valid Plücker coordinates p will have squaredError p == 0

That said, floating point makes a mockery of this claim, so you may want to use nearZero.

(><) :: forall a. Num a => Exp (Plucker a) -> Exp (Plucker a) -> Exp a infixl 5

This isn't the actual metric because this bilinear form gives rise to an isotropic quadratic space

plucker :: forall a. Num a => Exp (V4 a) -> Exp (V4 a) -> Exp (Plucker a)

Given a pair of points represented by homogeneous coordinates generate Plücker coordinates for the line through them, directed from the second towards the first.

plucker3D :: forall a. Num a => Exp (V3 a) -> Exp (V3 a) -> Exp (Plucker a)

Given a pair of 3D points, generate Plücker coordinates for the line through them, directed from the second towards the first.

Orphan instances

Additive Plucker

Methods

zero :: (Elt (Plucker a), Num a) => Exp (Plucker a)

(^+^) :: (Num a, Box Plucker a) => Exp (Plucker a) -> Exp (Plucker a) -> Exp (Plucker a)

(^-^) :: (Num a, Box Plucker a) => Exp (Plucker a) -> Exp (Plucker a) -> Exp (Plucker a)

lerp :: (Num a, Box Plucker a) => Exp a -> Exp (Plucker a) -> Exp (Plucker a) -> Exp (Plucker a)

Metric Plucker

Methods

dot :: (Num a, Box Plucker a) => Exp (Plucker a) -> Exp (Plucker a) -> Exp a

quadrance :: (Num a, Box Plucker a) => Exp (Plucker a) -> Exp a

qd :: (Num a, Box Plucker a) => Exp (Plucker a) -> Exp (Plucker a) -> Exp a

distance :: (Floating a, Box Plucker a) => Exp (Plucker a) -> Exp (Plucker a) -> Exp a

norm :: (Floating a, Box Plucker a) => Exp (Plucker a) -> Exp a

signorm :: (Floating a, Box Plucker a) => Exp (Plucker a) -> Exp (Plucker a)

cst a => IsProduct cst (Plucker a)

Associated Types

type ProdRepr (Plucker a) :: *

Methods

fromProd :: proxy cst -> Plucker a -> ProdRepr (Plucker a)

toProd :: proxy cst -> ProdRepr (Plucker a) -> Plucker a

prod :: proxy cst -> Plucker a -> ProdR cst (ProdRepr (Plucker a))

(Lift Exp a, Elt (Plain a)) => Lift Exp (Plucker a)

Associated Types

type Plain (Plucker a) :: *

Methods

lift :: Plucker a -> Exp (Plain (Plucker a))

Elt a => Unlift Exp (Plucker (Exp a))

Methods

unlift :: Exp (Plain (Plucker (Exp a))) -> Plucker (Exp a)

Elt a => Elt (Plucker a)

Methods

eltType :: Plucker a -> TupleType (EltRepr (Plucker a))

fromElt :: Plucker a -> EltRepr (Plucker a)

toElt :: EltRepr (Plucker a) -> Plucker a

(Elt a, Elt b) => Each (Exp (Plucker a)) (Exp (Plucker b)) (Exp a) (Exp b)

Methods

each :: Traversal (Exp (Plucker a)) (Exp (Plucker b)) (Exp a) (Exp b)